chore: restructured tests into separate packages
